Replace dataclasses with pydantic
