+++ pypolestar/auth.py
@@ -4,12 +4,11 @@ import hashlib
 import logging
 import os
 import re
-from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
-from typing import Self
 from urllib.parse import urljoin, urlparse
 
 import httpx
+from pydantic import BaseModel
 
 from .const import (
     HTTPX_TIMEOUT,
@@ -29,19 +28,10 @@ def b64urlencode(data: bytes) -> str:
     return base64.urlsafe_b64encode(data).decode().rstrip("=")
 
 
-@dataclass(frozen=True)
-class OidcConfiguration:
+class OidcConfiguration(BaseModel):
     issuer: str
     token_endpoint: str
     authorization_endpoint: str
-
-    @classmethod
-    def from_dict(cls, data: dict[str, str]) -> Self:
-        return cls(
-            issuer=data["issuer"],
-            token_endpoint=data["token_endpoint"],
-            authorization_endpoint=data["authorization_endpoint"],
-        )
 
 
 class PolestarAuth:
@@ -112,7 +102,7 @@ class PolestarAuth:
             raise PolestarAuthUnavailable(
                 message="Unable to get OIDC configuration", error_code=exc.response.status_code
             ) from exc
-        self.oidc_configuration = OidcConfiguration.from_dict(result.json())
+        self.oidc_configuration = OidcConfiguration.model_validate(result.json())
 
     def need_token_refresh(self) -> bool:
         """Return True if token needs refresh"""
