+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "grpcio>=1.75.1",
     "httpx>=0.28.1",
     "protobuf>=6.31.1",
+    "pydantic>=2.13.4",
 ]
 
 [project.urls]
